references/ADVANCED-transcription/speech-to-text/kyutai/fine-tuning/pyproject.toml
[project]
name = "fine-tuning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.1",
    "datasets>=4.0.0",
    "hf-transfer>=0.1.9",
    "huggingface-hub",
    "ipykernel>=6.30.1",
    "jiwer>=3.0.0",
    "julius>=0.2.7",
    "librosa>=0.11.0",
    "llvmlite>=0.40",
    "moshi==0.2.11",
    "numba>=0.57,<0.62",
    "peft>=0.17.1",
    "pytest>=8.4.2",
    "scipy>=1.16.1",
    "soundfile>=0.13.1",
    "torchaudio>=2.7.1",
    "torchcodec>=0.7.0",
    "transformers",
    "wandb>=0.18.0",
]
